refactor(functionality): Removed usage of YouTubeContext
