--- .env
@@ -1,2 +1,0 @@
-VITE_REACT_YOUTUBE_CLIENT_ID = "355803749408-kcs13sp2rj45tnebbro4lbbpctsl9p5v.apps.googleusercontent.com"
-VITE_REACT_YOUTUBE_API_KEY = "AIzaSyBe1ifDE1C3t0VIX1pdPPpDj3ElLWWN0Ag"
