+++ src/App.jsx
@@ -1,11 +1,12 @@
+import { useContext } from 'react'
+
 import Login from "./components/Login"
 import Main from "./components/Main-window"
-import {useContext} from 'react';
-import {Context} from "./Context"
-import './App.css'
+import { Context } from "./context/Context"
+import "./App.css"
 
 function App() {
-  const {accessToken} = useContext(Context);
+  const { accessToken } = useContext(Context);
 
   return (
     <>
