+++ src/main.jsx
@@ -1,15 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { ContextProvider } from "./Context"
-import { YouTubeContextProvider } from "./YouTubeContext"
+import { ContextProvider } from "./context/Context"
 import App from './App.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <YouTubeContextProvider>
-      <ContextProvider>
-        <App />
-      </ContextProvider>
-    </YouTubeContextProvider>
+    <ContextProvider>
+      <App />
+    </ContextProvider>
   </React.StrictMode>,
 )
